+++ src/globalStatusBar/clock/clock.js
