+++ src/globalStatusBar/clock/clock.test.js
