+++ src/globalStatusBar/contestSignature/contestName.js
