+++ src/globalStatusBar/contestSignature/contestSignature.js
