+++ src/globalStatusBar/globalStatusBar.js
