+++ src/globalStatusBar/globalStatusBar.test.js
