+++ src/globalStatusBar/login/libValidate.js
