+++ src/globalStatusBar/login/loginDialog.js
