+++ src/globalStatusBar/userSetting/passwordChange/passwordChangeDialog.js
