+++ src/globalStatusBar/userSetting/userSettingButton.js
