+++ src/globalStatusBar/userSetting/userSettingDialog.js
